requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
backtrader==1.9.78.123
binance==0.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
dateparser==1.2.1
fonttools==4.56.0
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
multidict==6.2.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
pycryptodome==3.22.0
pyparsing==3.2.1
python-binance==1.0.28
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
websockets==15.0.1
yarl==1.18.3
